AIGame/requirements.txt
flask>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
flask-cors>=4.0.0
